Move, rename, and update usages of folders and files
