+++ vinmonoholikern2/src/app/app.scss
@@ -1,4 +1,4 @@
-@import "../../assets/style/_shared.scss";
+@import "../assets/style/_shared.scss";
 
 #content {
   width: 80%;
