+++ vinmonoholikern2/src/components/Filter/FilterButton/index.tsx
@@ -1,110 +1,56 @@
-import React, {
-  ChangeEvent,
-  Dispatch,
-  FunctionComponent,
-  useState
-} from "react";
-import { SET_FILTER_ITEMS } from "../../reducers/constants";
-import {
-  listContainslist,
-  removeItem,
-  removeItems
-} from "../../utils/listFunctions";
-import FilterButton from "../filterButton/presentational";
-import { QueryAction } from "../../interfaces/ReducerInterfaces";
-import { Dropdown } from "../../interfaces/UtilityInterfaces";
-import "./filterDropdown.scss";
+import React, { Dispatch, FunctionComponent, useState } from "react";
+import productDropdownList from "../../../utils/CategoryTypes";
+import SubCategoryDropdown from "../SubCategoryDropdown";
+import { QueryAction } from "../../../interfaces/ReducerInterfaces";
+import { chevronDown, chevronUp } from "../constants";
+import "./FilterButton.scss";
 
 interface Props {
-  dropdown: Dropdown;
-  dispatch: Dispatch<QueryAction>;
+  dispatchQuery: Dispatch<QueryAction>;
   filterItems: string[];
 }
 
-const chevronUpPath = "./images/chevron_up.svg";
-const chevronDownPath = "./images/chevron_down.svg";
+const FilterButton: FunctionComponent<Props> = (props: Props): JSX.Element => {
+  const { dispatchQuery, filterItems } = props;
+  const [showFilters, setShow] = useState(false);
+  const toggleShowFilter = (): void => setShow(!showFilters);
+  const chevronPath = showFilters ? chevronUp : chevronDown;
+  const hideFilters = showFilters ? undefined : { display: "none" };
+  const toggleFilterId = showFilters
+    ? "filter-head-active"
+    : "filter-head-unactive";
 
-const FilterDropdown: FunctionComponent<Props> = (props: Props): JSX.Element => {
-  const { dropdown, dispatch, filterItems } = props;
-  const [showDropdown, toggleDropdown] = useState(false);
-
-  const itemOnClick = (active: boolean, category: string): void => {
-    let payload;
-    if (active) {
-      payload = [...filterItems, category];
-      dispatch({ type: SET_FILTER_ITEMS, payload });
-    } else {
-      payload = removeItem(filterItems, category);
-      dispatch({ type: SET_FILTER_ITEMS, payload });
-    }
-  };
-
-  const selectedCategories = filterItems.filter((item): boolean => dropdown.subCategories.includes(item));
-
-  const checkboxOnChange = (event: ChangeEvent<HTMLInputElement>): void => {
-    const { checked } = event.target;
-    const { subCategories } = dropdown;
-    let processedItems;
-    if (checked) {
-      processedItems = subCategories.filter((item): boolean =>
-        subCategories.includes(item)
-      );
-      dispatch({
-        type: SET_FILTER_ITEMS,
-        payload: [...filterItems, ...processedItems]
-      });
-    } else {
-      processedItems = removeItems(filterItems, subCategories);
-      dispatch({ type: SET_FILTER_ITEMS, payload: [...processedItems] });
-    }
-  };
-
-  const isChecked = listContainslist(dropdown.subCategories, filterItems);
-  const chevron = showDropdown ? chevronUpPath : chevronDownPath;
-  const handleKeyPress = (event: React.KeyboardEvent): void => {
-    if (event.key === "Enter") toggleDropdown(!showDropdown);
-  }
-
-  const categoryButtons = dropdown.subCategories.map((category): JSX.Element => (
-    <FilterButton
-      active={selectedCategories.includes(category)}
-      key={category}
-      name={category}
-      onClick={itemOnClick}
+  const toggleButtons = productDropdownList.map((dropdown): JSX.Element => (
+    <SubCategoryDropdown
+      key={dropdown.name}
+      dropdown={dropdown}
+      filterItems={filterItems}
+      dispatch={dispatchQuery}
     />
   ));
-  const htmlLabel = `checkbox-${dropdown.name}`;
+
+  const handleKeyPress = (event: React.KeyboardEvent): void => {
+    if (event.key === "Enter") toggleShowFilter();
+  }
+
   return (
-    <div>
-      <span className="dropdown-head">
-        <span className="dropdown-checkbox">
-          <label htmlFor={htmlLabel}>
-            <input
-              type="checkbox"
-              onChange={checkboxOnChange}
-              checked={isChecked}
-              id={htmlLabel}
-            />
-          </label>
-        </span>
-        <span
-          className="dropdown-title"
-          onClick={(): void => toggleDropdown(!showDropdown)}
-          onKeyPress={handleKeyPress}
-          role="button"
-          tabIndex={-1}
-        >
-          <span className="dropdown-category">{dropdown.name}</span>
-          <img
-            className="dropdown-chevron"
-            src={chevron}
-            alt="Open/Close sub categories"
-          />
-        </span>
-      </span>
-      {showDropdown ? categoryButtons : null}
+    <div id="filter-container">
+      <div
+        className="filter-head"
+        id={toggleFilterId}
+        onClick={toggleShowFilter}
+        onKeyPress={handleKeyPress}
+        role="button"
+        tabIndex={-1}
+      >
+        <h2>Varetype filter</h2>
+        <img src={chevronPath} alt="open/close filter" />
+      </div>
+      <div id="filter-items" style={hideFilters}>
+        {toggleButtons}
+      </div>
     </div>
   );
 };
 
-export default FilterDropdown;
+export default FilterButton;
