+++ vinmonoholikern2/src/components/Filter/SubCategoryButton/SubCategoryButton.scss
@@ -1,4 +1,4 @@
-@import "../../assets/style/shared";
+@import "../../../assets/style/shared";
 $active-color: #cbef53;
 @mixin toggle-shadow {
   -webkit-box-shadow: 8px 10px 13px -9px rgba(0,0,0,0.75);
