+++ vinmonoholikern2/src/components/Filter/constants.ts
@@ -1,0 +1,2 @@
+export const chevronUp = "./images/chevron_up.svg";
+export const chevronDown = "./images/chevron_down.svg";
