+++ vinmonoholikern2/src/components/LoadingBar/LoadingBar.scss
