+++ vinmonoholikern2/src/components/ProductTable/Pageable/Pageable.scss
@@ -1,4 +1,4 @@
-@import "../../assets/style/_shared.scss";
+@import "../../../assets/style/_shared.scss";
 
 #pageable-container {
   display: grid;
