+++ vinmonoholikern2/src/components/ProductTable/TableHeader/TableHeader.scss
@@ -1,4 +1,4 @@
-@import '../../assets/style/_shared.scss';
+@import '../../../assets/style/_shared.scss';
 
 #item-table-head {
   background-color: $primary-color;
