+++ vinmonoholikern2/src/components/ProductTable/TableRow/TableRow.scss
@@ -1,5 +1,5 @@
-@import '../../assets/style/_shared.scss';
-@import '../itemTable/itemTable.scss';
+@import '../../../assets/style/_shared.scss';
+@import '../Table/Table.scss';
 
 #item-table-body {
   // background-color: $primary-color;
