+++ vinmonoholikern2/src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/app/App";
+import App from "./app/App";
 import "./index.scss";
 import * as serviceWorker from "./serviceWorker";
